Remove obsolete UploadModal
