+++ di-billing-app/apps/web/src/pages/DashboardPage.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+
+export function DashboardPage() {
+  return (
+    <div className="p-6">
+      <h1 className="text-2xl font-semibold text-slate-100">Dashboard</h1>
+      <p className="mt-2 text-slate-400">Welcome to the DI Billing App. Analytics and summaries will be shown here.</p>
+    </div>
+  );
+}
