+++ di-billing-app/apps/web/src/pages/MappingsPage.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+
+export function MappingsPage() {
+  return (
+    <div className="p-6">
+      <h1 className="text-2xl font-semibold text-slate-100">Product Mappings</h1>
+      <p className="mt-2 text-slate-400">An editable data grid for managing Salesforce Product Code to GM Program mappings will be here.</p>
+    </div>
+  );
+}
