+++ di-billing-app/apps/web/src/pages/ReportsPage.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+
+export function ReportsPage() {
+  return (
+    <div className="p-6">
+      <h1 className="text-2xl font-semibold text-slate-100">Reports</h1>
+      <p className="mt-2 text-slate-400">A list of saved and exported discrepancy reports will be available here.</p>
+    </div>
+  );
+}
